requirements.txt
requests >= 2.27.1
parsel >= 1.6.0
pyyaml >= 6.0